# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Documentation of our awesome API.",
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/v1",
  "paths": {
...
    "/donor/donors": {
      "get": {
        "description": "You receive an list of donors",
        "tags": [
          "Donor"
        ],
        "operationId": "ListDonors",
        "responses": {
          "200": {
            "$ref": "#/responses/donorListResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
    "/oauth/google/login": {
      "get": {
        "description": "Need to send request for google account check",
        "tags": [
          "OAuth"
        ],
        "operationId": "OauthGoogleLogin",
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "401": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
    "/refresh-token": {
      "get": {
        "description": "Send request within refresh-token to get new access-token",
        "tags": [
          "Auth"
        ],
        "operationId": "RefreshToken",
        "responses": {
          "200": {
            "$ref": "#/responses/genericResponse"
          },
          "401": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
    "/user/current": {
      "get": {
        "description": "You receive an user with id from JWT",
        "tags": [
          "User"
        ],
        "operationId": "CurrentUser",
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
    "/user/users": {
      "get": {
        "description": "You receive an list of users",
        "tags": [
          "User"
        ],
        "operationId": "ListUsers",
        "responses": {
          "200": {
            "$ref": "#/responses/userListResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
    "/user/{user_id}": {
      "get": {
        "description": "You receive an list of users",
        "tags": [
          "User"
        ],
        "operationId": "GetUserById",
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      },
      "put": {
        "description": "You receive an list of users",
        "tags": [
          "User"
        ],
        "operationId": "UpdateUserById",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "addres": {
                  "type": "string",
                  "x-go-name": "Address"
                },
                "name": {
                  "type": "string",
                  "x-go-name": "Name"
                },
                "username": {
                  "type": "string",
                  "x-go-name": "Username"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/userResponse"
          },
          "500": {
            "$ref": "#/responses/genericResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GenericResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "x-go-name": "Data"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "boolean",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "_/home/dzemil/Projects/golang/risc_monolith/server/utils"
    },
...
    "showDonor": {
      "type": "object",
      "properties": {
        "donor_address": {
          "type": "string",
          "x-go-name": "DonorAddress"
        },
        "donor_blood_type": {
          "type": "string",
          "x-go-name": "DonorBloodType"
        },
        "donor_blood_type_num": {
          "type": "integer",
          "format": "int16",
          "x-go-name": "DonorBloodTypeNum"
        },
        "donor_last_donation": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DonorLastDonationDate"
        },
        "donor_name": {
          "type": "string",
          "x-go-name": "DonorName"
        },
        "donor_phone_number": {
          "type": "string",
          "x-go-name": "DonorPhoneNumber"
        },
        "donor_surname": {
          "type": "string",
          "x-go-name": "DonorSurname"
        },
        "donor_unique_identification_number": {
          "type": "string",
          "x-go-name": "DonorUniqueIdentificationNumber"
        }
      },
      "x-go-package": "_/home/dzemil/Projects/golang/risc_monolith/server/utils"
    },
    "showUsers": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id"
        },
        "isverified": {
          "type": "boolean",
          "x-go-name": "Isverified"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "role": {
          "type": "string",
          "x-go-name": "Role"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "_/home/dzemil/Projects/golang/risc_monolith/server/utils"
    },
...
  },
  "responses": {
...
    "donorListResponse": {
      "description": "List of donors with basic info",
      "schema": {
        "type": "object",
        "properties": {
          "Donors": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/showDonor"
            }
          }
        }
      }
    },
    "genericResponse": {
      "description": "This text will appear as description of your response body.",
      "schema": {
        "$ref": "#/definitions/GenericResponse"
      }
    },
...
    "userListResponse": {
      "description": "List of users with basic info",
      "schema": {
        "type": "object",
        "properties": {
          "Users": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/showUsers"
            }
          }
        }
      }
    },
    "userResponse": {
      "description": "User with basic info",
      "schema": {
        "type": "object",
        "properties": {
          "User": {
            "$ref": "#/definitions/showUsers"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}